--- MC1GD/Assets.xcassets/beruang-welcome.imageset/Contents.json
@@ -1,21 +1,0 @@
-{
-  "images" : [
-    {
-      "filename" : "Beruang welcome.png",
-      "idiom" : "universal",
-      "scale" : "1x"
-    },
-    {
-      "idiom" : "universal",
-      "scale" : "2x"
-    },
-    {
-      "idiom" : "universal",
-      "scale" : "3x"
-    }
-  ],
-  "info" : {
-    "author" : "xcode",
-    "version" : 1
-  }
-}
