+++ MC1GD/Views/RootView.swift
@@ -24,9 +24,8 @@ struct RootView: View {
                 if selectedTab == .summary {
                     SummaryView(todayDateComponent: $todayDateComponent, data: $data)
                 }else if selectedTab == .expense {
-                    ExpenseView(todayDateComponent: $todayDateComponent)
-                }else if selectedTab == .tabungan {
-                    SavingsView(todayDateComponent: $todayDateComponent)
+                    //                ExpenseView(todayDateComponent: $todayDateComponent)
+                    ExpenseViewTemp(todayDateComponent: $todayDateComponent)
                 }
                 Spacer()
